# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Persistent shared memory infrastructure for AI agents. Stores observations in PostgreSQL + pgvector, exposes MCP tools via HTTP/SSE.",
        "title": "Engram API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:37777",
    "basePath": "/",
    "paths": {
...
        "/api/backfill/session": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Accepts raw JSONL session content, runs server-side LLM extraction, and stores resulting observations with semantic deduplication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backfill"
                ],
                "summary": "Backfill session with LLM extraction",
                "parameters": [
                    {
                        "description": "Session content and metadata",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_worker.BackfillSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_worker.BackfillSessionResponse"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "LLM not configured or store not initialized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/backfill/status": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns status of all active backfill runs including stored/skipped/error counts per run.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backfill"
                ],
                "summary": "Get backfill status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_worker.BackfillStatus"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_worker.BackfillSessionRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Content is the raw JSONL session data.",
                    "type": "string"
                },
                "project": {
                    "description": "Project overrides the project path from session metadata. Empty = use parsed value.",
                    "type": "string"
                },
                "run_id": {
                    "description": "RunID groups observations from the same backfill run.",
                    "type": "string"
                },
                "session_id": {
                    "description": "SessionID identifies the source session (e.g. UUID from filename).",
                    "type": "string"
                }
            }
        },
        "internal_worker.BackfillSessionResponse": {
            "type": "object",
            "properties": {
                "deprecated": {
                    "description": "Explains that backfill persistence was removed in v5.",
                    "type": "string"
                },
                "errors": {
                    "type": "integer"
                },
                "skipped": {
                    "type": "integer"
                },
                "stored": {
                    "type": "integer"
                }
            }
        },
        "internal_worker.BackfillStatus": {
            "type": "object",
            "properties": {
                "active_runs": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/internal_worker.RunInfo"
                    }
                },
                "total_observations": {
                    "type": "integer"
                },
                "total_runs": {
                    "type": "integer"
                }
            }
        },
...
        "internal_worker.RunInfo": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "integer"
                },
                "run_id": {
                    "type": "string"
                },
                "sessions": {
                    "type": "integer"
                },
                "skipped": {
                    "type": "integer"
                },
                "stored": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-Auth-Token",
            "in": "header"
        }
    }
}